+++ man-roxygen/examples-description.R
@@ -1,0 +1,10 @@
+#' @examples
+#' library(tidyhydro)
+#'
+#' <%=fn %>
+#' 
+#' # Supply truth as bare column names
+#' <%=fn %>(avacha, obs)
+#'
+#' # Or as numeric vectors
+#' <%=fn %>_vec(avacha$obs)