+++ man-roxygen/examples-numeric.R
@@ -1,6 +1,5 @@
 #' @examples
 #' library(tidyhydro)
-#' data(avacha)
 #'
 #' # Supply truth and predictions as bare column names
 #' <%=fn %>(avacha, obs, sim)