+++ tests/testthat/test-hydrogof.R
@@ -1,6 +1,9 @@
 # Property-based testing
 # https://www.etiennebacher.com/posts/2024-10-01-using-property-testing-in-r
 
+# Tests to ensure an accordance in results between hydroGOF and
+# tidyhydro packages.
+
 options(
   quickcheck.tests = 20L,
   quickcheck.shrinks = 10L,
@@ -16,14 +19,14 @@ test_that("nse", {
   # same results as hydroGOF with both NA values present and without
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -38,14 +41,14 @@ test_that("nse", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -65,14 +68,14 @@ test_that("kge", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -87,14 +90,14 @@ test_that("kge", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -115,14 +118,14 @@ test_that("kge2012", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -142,14 +145,14 @@ test_that("kge2012", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -167,6 +170,65 @@ test_that("kge2012", {
   )
 })
 
+test_that("kgelog", {
+  skip_if_not_installed("quickcheck")
+  skip_if_not_installed("hydroGOF")
+
+  # With NA
+  quickcheck::for_all(
+    obs = quickcheck::double_bounded(
+      left = 1,
+      right = 2000,
+      len = 50,
+      any_na = TRUE
+    ),
+    sim = quickcheck::double_bounded(
+      left = 1,
+      right = 2000,
+      len = 50,
+      any_na = TRUE
+    ),
+    property = function(obs, sim) {
+      new <- kgelog_vec(truth = obs, estimate = sim, na_rm = TRUE)
+      old <- hydroGOF::KGE(
+        sim = log10(sim),
+        obs = log10(obs),
+        na.rm = TRUE,
+        method = "2012"
+      )
+
+      expect_equal(new, old)
+    }
+  )
+
+  # Without NA
+  quickcheck::for_all(
+    obs = quickcheck::double_bounded(
+      left = 1,
+      right = 2000,
+      len = 50,
+      any_na = FALSE
+    ),
+    sim = quickcheck::double_bounded(
+      left = 1,
+      right = 2000,
+      len = 50,
+      any_na = FALSE
+    ),
+    property = function(obs, sim) {
+      new <- kgelog_vec(truth = obs, estimate = sim, na_rm = TRUE)
+      old <- hydroGOF::KGE(
+        sim = log10(sim),
+        obs = log10(obs),
+        na.rm = TRUE,
+        method = "2012"
+      )
+
+      expect_equal(new, old)
+    }
+  )
+})
+
 test_that("pbias", {
   skip_if_not_installed("quickcheck")
   skip_if_not_installed("hydroGOF")
@@ -174,14 +236,14 @@ test_that("pbias", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -196,14 +258,14 @@ test_that("pbias", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
@@ -223,14 +285,14 @@ test_that("mse", {
   # With NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = TRUE
     ),
@@ -245,20 +307,69 @@ test_that("mse", {
   # Without NA
   quickcheck::for_all(
     obs = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     sim = quickcheck::double_bounded(
-      left = -1000,
-      right = 1000,
+      left = -2000,
+      right = 2000,
       len = 50,
       any_na = FALSE
     ),
     property = function(obs, sim) {
       new <- mse_vec(truth = obs, estimate = sim, na_rm = TRUE)
       old <- hydroGOF::mse(sim = sim, obs = obs, na.rm = TRUE)
+
+      expect_equal(new, old)
+    }
+  )
+})
+
+test_that("rmse", {
+  skip_if_not_installed("quickcheck")
+  skip_if_not_installed("hydroGOF")
+
+  # With NA
+  quickcheck::for_all(
+    obs = quickcheck::double_bounded(
+      left = -2000,
+      right = 2000,
+      len = 50,
+      any_na = TRUE
+    ),
+    sim = quickcheck::double_bounded(
+      left = -2000,
+      right = 2000,
+      len = 50,
+      any_na = TRUE
+    ),
+    property = function(obs, sim) {
+      new <- rmse_vec(truth = obs, estimate = sim, na_rm = TRUE)
+      old <- hydroGOF::rmse(sim = sim, obs = obs, na.rm = TRUE)
+
+      expect_equal(new, old)
+    }
+  )
+
+  # Without NA
+  quickcheck::for_all(
+    obs = quickcheck::double_bounded(
+      left = -2000,
+      right = 2000,
+      len = 50,
+      any_na = FALSE
+    ),
+    sim = quickcheck::double_bounded(
+      left = -2000,
+      right = 2000,
+      len = 50,
+      any_na = FALSE
+    ),
+    property = function(obs, sim) {
+      new <- rmse_vec(truth = obs, estimate = sim, na_rm = TRUE)
+      old <- hydroGOF::rmse(sim = sim, obs = obs, na.rm = TRUE)
 
       expect_equal(new, old)
     }